refactor: 💡 Create network namespace
Creates a network namespace that contains all previous middleware
methods as well as some from root

- sdk.getLastestBlock -> sdk.network.getLatestBlock
- sdk.getNetworkVersion -> sdk.network.getVersion
- sdk.getNetworkProperties -> sdk.network.getProperties
- sdk.getTransactionFees -> sdk.network.getTransactionFees
- sdk.getTreasuryAccount -> sdk.network.getTreasuryAccount
- sdk.getTreasuryBalance -> sdk.network.getTreasuryBalance
- sdk.transferPolyx -> sdk.network.transferPolyx
- sdk.middleware.getEventByIndexedArgs -> sdk.network.getEventByIndexedArgs
- sdk.middleware.getEventsByIndexedArgs -> sdk.network.getEventsByIndexedArgs
- sdk.middleware.getTransactionByHash -> sdk.network.getTransactionByHash

BREAKING CHANGE: 🧨 Moved methods
