+++ src/Polymesh.ts
@@ -5,10 +5,8 @@ import { ApolloClient } from 'apollo-client';
 import { ApolloLink } from 'apollo-link';
 import { setContext } from 'apollo-link-context';
 import { HttpLink } from 'apollo-link-http';
-import BigNumber from 'bignumber.js';
 import fetch from 'cross-fetch';
 import schema from 'polymesh-types/schema';
-import { TxTag } from 'polymesh-types/types';
 import { satisfies } from 'semver';
 import { w3cwebsocket as W3CWebSocket } from 'websocket';
 import WebSocketAsPromised from 'websocket-as-promised';
@@ -24,8 +22,6 @@ import {
   RegisterIdentityParams,
   SecurityToken,
   TickerReservation,
-  transferPolyx,
-  TransferPolyxParams,
 } from '~/internal';
 import { heartbeat } from '~/middleware/queries';
 import { Settlements } from '~/Settlements';
@@ -34,7 +30,6 @@ import {
   CommonKeyring,
   ErrorCode,
   MiddlewareConfig,
-  NetworkProperties,
   ProcedureMethod,
   SubCallback,
   TickerReservationStatus,
@@ -42,25 +37,18 @@ import {
   UnsubCallback,
 } from '~/types';
 import {
-  moduleAddressToString,
   signerToString,
   stringToIdentityId,
   stringToTicker,
-  textToString,
   tickerToString,
-  u32ToBigNumber,
 } from '~/utils/conversion';
 import { createProcedureMethod, getDid, isPrintableAscii } from '~/utils/internal';
 
 import { Claims } from './Claims';
 import { CurrentIdentity } from './CurrentIdentity';
 // import { Governance } from './Governance';
-import { Middleware } from './Middleware';
-import {
-  SUPPORTED_VERSION_RANGE,
-  SYSTEM_VERSION_RPC_CALL,
-  TREASURY_MODULE_ADDRESS,
-} from './utils/constants';
+import { Network } from './Network';
+import { SUPPORTED_VERSION_RANGE, SYSTEM_VERSION_RPC_CALL } from './utils/constants';
 
 interface ConnectParamsBase {
   nodeUrl: string;
@@ -79,7 +67,7 @@ export class Polymesh {
   // NOTE uncomment in Governance v2 upgrade
   // public governance: Governance;
   public claims: Claims;
-  public middleware: Middleware;
+  public network: Network;
   public settlements: Settlements;
   public currentIdentity: CurrentIdentity;
 
@@ -92,14 +80,9 @@ export class Polymesh {
     // NOTE uncomment in Governance v2 upgrade
     // this.governance = new Governance(context);
     this.claims = new Claims(context);
-    this.middleware = new Middleware(context);
+    this.network = new Network(context);
     this.settlements = new Settlements(context);
     this.currentIdentity = new CurrentIdentity(context);
-
-    this.transferPolyx = createProcedureMethod(
-      { getProcedureAndArgs: args => [transferPolyx, args] },
-      context
-    );
 
     this.registerIdentity = createProcedureMethod(
       { getProcedureAndArgs: args => [registerIdentity, args] },
@@ -293,11 +276,6 @@ export class Polymesh {
 
     return new Polymesh(context);
   }
-
-  /**
-   * Transfer an amount of POLYX to a specified Account
-   */
-  public transferPolyx: ProcedureMethod<TransferPolyxParams, void>;
 
   /**
    * Get the free/locked POLYX balance of an Account
@@ -517,26 +495,6 @@ export class Polymesh {
   }
 
   /**
-   * Retrieve the protocol fees associated with running a specific transaction
-   *
-   * @param args.tag - transaction tag (i.e. TxTags.asset.CreateAsset or "asset.createAsset")
-   */
-  public getTransactionFees(args: { tag: TxTag }): Promise<BigNumber> {
-    return this.context.getTransactionFees(args.tag);
-  }
-
-  /**
-   * Get the treasury wallet address
-   */
-  public getTreasuryAccount(): Account {
-    const { context } = this;
-    return new Account(
-      { address: moduleAddressToString(TREASURY_MODULE_ADDRESS, context) },
-      context
-    );
-  }
-
-  /**
    * Handle connection errors
    *
    * @returns an unsubscribe callback
@@ -626,52 +584,6 @@ export class Polymesh {
   }
 
   /**
-   * Retrieve information for the current network
-   */
-  public async getNetworkProperties(): Promise<NetworkProperties> {
-    const {
-      context: {
-        polymeshApi: {
-          runtimeVersion: { specVersion },
-          rpc: {
-            system: { chain },
-          },
-        },
-      },
-    } = this;
-    const name = await chain();
-
-    return {
-      name: textToString(name),
-      version: u32ToBigNumber(specVersion).toNumber(),
-    };
-  }
-
-  /**
-   * Get the Treasury POLYX balance
-   *
-   * @note can be subscribed to
-   */
-  public getTreasuryBalance(): Promise<BigNumber>;
-  public getTreasuryBalance(callback: SubCallback<BigNumber>): Promise<UnsubCallback>;
-
-  // eslint-disable-next-line require-jsdoc
-  public async getTreasuryBalance(
-    callback?: SubCallback<BigNumber>
-  ): Promise<BigNumber | UnsubCallback> {
-    const account = this.getTreasuryAccount();
-
-    if (callback) {
-      return account.getBalance(({ free: freeBalance }) => {
-        callback(freeBalance);
-      });
-    }
-
-    const { free } = await account.getBalance();
-    return free;
-  }
-
-  /**
    * Register an Identity
    *
    * @note must be a CDD provider
@@ -683,13 +595,6 @@ export class Polymesh {
    *   - Customer Due Diligence Provider
    */
   public registerIdentity: ProcedureMethod<RegisterIdentityParams, Identity>;
-
-  /**
-   * Retrieve the number of the latest block in the chain
-   */
-  public getLatestBlock(): Promise<BigNumber> {
-    return this.context.getLatestBlock();
-  }
 
   /**
    * Disconnect the client and close all open connections and subscriptions
@@ -745,13 +650,6 @@ export class Polymesh {
    */
   public setSigner(signer: string | Account): void {
     this.context.setPair(signerToString(signer));
-  }
-
-  /**
-   * Fetch the current network version (i.e. 3.1.0)
-   */
-  public async getNetworkVersion(): Promise<string> {
-    return this.context.getNetworkVersion();
   }
 
   // TODO @monitz87: remove when the dApp team no longer needs it
