+++ src/__tests__/Polymesh.ts
@@ -3,7 +3,6 @@ import { Signer as PolkadotSigner } from '@polkadot/api/types';
 import { ApolloLink, GraphQLRequest } from 'apollo-link';
 import * as apolloLinkContextModule from 'apollo-link-context';
 import BigNumber from 'bignumber.js';
-import { TxTags } from 'polymesh-types/types';
 import semver from 'semver';
 import sinon from 'sinon';
 
@@ -787,33 +786,6 @@ describe('Polymesh Class', () => {
     });
   });
 
-  describe('method: getTransactionFees', () => {
-    test('should return the fees associated to the supplied transaction', async () => {
-      dsMockUtils.configureMocks({ contextOptions: { transactionFee: new BigNumber(500) } });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const fee = await polymesh.getTransactionFees({ tag: TxTags.asset.CreateAsset });
-
-      expect(fee).toEqual(new BigNumber(500));
-    });
-  });
-
-  describe('method: getTreasuryAccount', () => {
-    test('should return the treasury account', async () => {
-      const treasuryAddress = '5EYCAe5ijAx5xEfZdpCna3grUpY1M9M5vLUH5vpmwV1EnaYR';
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      expect(polymesh.getTreasuryAccount().address).toEqual(treasuryAddress);
-    });
-  });
-
   describe('method: getSecurityTokens', () => {
     beforeAll(() => {
       sinon.stub(utilsConversionModule, 'signerValueToSignatory');
@@ -911,33 +883,6 @@ describe('Polymesh Class', () => {
     });
   });
 
-  describe('method: transferPolyx', () => {
-    test('should prepare the procedure with the correct arguments and context, and return the resulting transaction queue', async () => {
-      const context = dsMockUtils.getContextInstance();
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      const args = {
-        to: 'someAccount',
-        amount: new BigNumber(50),
-      };
-
-      const expectedQueue = ('' as unknown) as TransactionQueue<void>;
-
-      procedureMockUtils
-        .getPrepareStub()
-        .withArgs({ args, transformer: undefined }, context)
-        .resolves(expectedQueue);
-
-      const queue = await polymesh.transferPolyx(args);
-
-      expect(queue).toBe(expectedQueue);
-    });
-  });
-
   describe('method: getSecurityToken', () => {
     test('should return a specific Security Token', async () => {
       const ticker = 'TEST';
@@ -963,68 +908,6 @@ describe('Polymesh Class', () => {
       return expect(polymesh.getSecurityToken({ ticker })).rejects.toThrow(
         `There is no Security Token with ticker "${ticker}"`
       );
-    });
-  });
-
-  describe('method: getNetworkProperties', () => {
-    test('should return current network information', async () => {
-      const name = 'someName';
-      const version = 1;
-      const fakeResult = {
-        name,
-        version,
-      };
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-      });
-
-      dsMockUtils.setRuntimeVersion({ specVersion: dsMockUtils.createMockU32(version) });
-      dsMockUtils.createRpcStub('system', 'chain').resolves(dsMockUtils.createMockText(name));
-
-      const result = await polymesh.getNetworkProperties();
-      expect(result).toEqual(fakeResult);
-    });
-  });
-
-  describe('method: getTreasuryBalance', () => {
-    let fakeBalance: AccountBalance;
-
-    beforeAll(() => {
-      fakeBalance = {
-        free: new BigNumber(500000),
-        locked: new BigNumber(0),
-        total: new BigNumber(500000),
-      };
-      entityMockUtils.configureMocks({ accountOptions: { getBalance: fakeBalance } });
-    });
-
-    test('should return the POLYX balance of the treasury account', async () => {
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-      });
-
-      const result = await polymesh.getTreasuryBalance();
-      expect(result).toEqual(fakeBalance.free);
-    });
-
-    test('should allow subscription', async () => {
-      const unsubCallback = 'unsubCallback';
-
-      entityMockUtils.getAccountInstance().getBalance.callsFake(async cbFunc => {
-        cbFunc(fakeBalance);
-        return unsubCallback;
-      });
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-      });
-
-      const callback = sinon.stub();
-      const result = await polymesh.getTreasuryBalance(callback);
-      expect(result).toEqual(unsubCallback);
-      sinon.assert.calledWithExactly(callback, fakeBalance.free);
     });
   });
 
@@ -1102,28 +985,6 @@ describe('Polymesh Class', () => {
     });
   });
 
-  describe('method: getLatestBlock', () => {
-    test('should return the latest block number', async () => {
-      const blockNumber = new BigNumber(100);
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true, latestBlock: blockNumber } });
-      dsMockUtils.createApolloQueryStub(heartbeat(), true);
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-        middleware: {
-          link: 'someLink',
-          key: 'someKey',
-        },
-      });
-
-      const result = await polymesh.getLatestBlock();
-
-      expect(result).toEqual(blockNumber);
-    });
-  });
-
   describe('method: disconnect', () => {
     test('should call the underlying disconnect function', async () => {
       const polymesh = await Polymesh.connect({
@@ -1137,28 +998,6 @@ describe('Polymesh Class', () => {
 
       await polymesh.disconnect();
       sinon.assert.calledOnce(dsMockUtils.getContextInstance().disconnect);
-    });
-  });
-
-  describe('method: getNetworkVersion', () => {
-    test('should return the network version', async () => {
-      const networkVersion = '1.0.0';
-
-      dsMockUtils.configureMocks({ contextOptions: { withSeed: true, networkVersion } });
-      dsMockUtils.createApolloQueryStub(heartbeat(), true);
-
-      const polymesh = await Polymesh.connect({
-        nodeUrl: 'wss://some.url',
-        accountUri: '//uri',
-        middleware: {
-          link: 'someLink',
-          key: 'someKey',
-        },
-      });
-
-      const result = await polymesh.getNetworkVersion();
-
-      expect(result).toEqual(networkVersion);
     });
   });
